refactor, moved logic into util file and pres into selection control
